refactor: move profile to sidebar
